requirements.txt
autogen-agentchat==0.4.8
autogen-ext[openai,azure,docker]==0.4.8
numpy
pandas
matplotlib
scikit-learn
python-dotenv
rich
